chore(swipe): split demo
