+++ packages/nutui-taro-demo/src/feedback/pages/swipe/basic.vue
@@ -1,0 +1,8 @@
+<template>
+  <nut-swipe>
+    <nut-cell round-radius="0" title="Swipe Left" />
+    <template #right>
+      <nut-button shape="square" style="height: 100%" type="danger">Button</nut-button>
+    </template>
+  </nut-swipe>
+</template>
