+++ packages/nutui-taro-demo/src/feedback/pages/swipe/disabled.vue
@@ -1,0 +1,8 @@
+<template>
+  <nut-swipe disabled>
+    <nut-cell round-radius="0" title="Disabled" />
+    <template #right>
+      <nut-button shape="square" style="height: 100%" type="danger">Button</nut-button>
+    </template>
+  </nut-swipe>
+</template>
