+++ src/packages/__VUE/swipe/demo/index.vue
@@ -21,7 +21,7 @@
 </template>
 
 <script setup lang="ts">
-import { useTranslate } from '../../../utils';
+import { useTranslate } from '@/sites/utils';
 import Basic from './basic.vue';
 import Disabled from './disabled.vue';
 import Direction from './direction.vue';
